+++ CHANGELOG.md
@@ -18,6 +18,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Throw `SerializationError` over client safe `Error` when failing to serialize leaf types
 - Move debug entries in errors under `extensions` key
 - Use native PHP types for `Schema` and `SchemaConfig`
+- Always throw `RequestError` with useful message when clients provide an invalid JSON body
 
 ### Added